Split Admin Panel into Admin Main Panel and Admin User Panel.
Main panel now will contain only Main buttons. Choosing users now will be processed with Admin User Panel.

Rename Settings Panel to Admin Settings Panel.
