+++ src/main/java/world/bentobox/addons/biomes/panel/AdminMainPanel.java
@@ -19,9 +19,9 @@ import world.bentobox.bentobox.api.user.User;
  * This class creates AdminPanel GUI that allows to change user biome, edit, add or remove biome and import
  * biomes.
  */
-public class AdminPanel
+public class AdminMainPanel
 {
-	public AdminPanel(BiomesAddon addon, World world, User user)
+	public AdminMainPanel(BiomesAddon addon, World world, User user)
 	{
 		this.addon = addon;
 		this.world = world;
@@ -45,7 +45,9 @@ public class AdminPanel
 			name(this.player.getTranslation("biomes.admin.buttons.change")).
 			icon(Material.LEVER).
 			clickHandler((panel, user, clickType, slot) -> {
-				this.createPlayerListMenu(0);
+				new AdminUserPanel(this.addon,
+					this.world,
+					this.player);
 				return true;
 			}).build());
 
@@ -96,104 +98,11 @@ public class AdminPanel
 			name(this.player.getTranslation("biomes.admin.buttons.settings")).
 			icon(Material.ENCHANTING_TABLE).
 			clickHandler((panel, user, clickType, slot) -> {
-				new SettingsPanel(this.addon,
+				new AdminSettingsPanel(this.addon,
 					this.world,
 					this.player);
 				return true;
 			}).build());
-
-		panelBuilder.build();
-	}
-
-
-	/**
-	 * This method creates panel with all players which has islands. On player head click it will open
-	 * BiomesPanel in ADMIN mode, that will allow to change selected user biome.
-	 * @param pageIndex Index of page.
-	 */
-	private void createPlayerListMenu(int pageIndex)
-	{
-		PanelBuilder panelBuilder = new PanelBuilder().user(this.player).name(
-			this.player.getTranslation("biomes.admin.gui-title"));
-
-		List<User> activeUsers = new ArrayList<>(this.world.getPlayers().size());
-
-		// Collect all players to create correct menu.
-		int activePlayersCount = 0;
-
-		for (Player player : this.world.getPlayers())
-		{
-			if (this.addon.getIslands().hasIsland(this.world, player.getUniqueId()))
-			{
-				activeUsers.add(this.addon.getPlayers().getUser(player.getUniqueId()));
-				activePlayersCount++;
-			}
-		}
-
-		// Normalize page INDEX.
-		if (pageIndex < 0)
-		{
-			pageIndex = 0;
-		}
-		else if (pageIndex > (activePlayersCount / MAX_PLAYERS_PER_PAGE))
-		{
-			pageIndex = activePlayersCount / MAX_PLAYERS_PER_PAGE;
-		}
-
-		int playerIndex = MAX_PLAYERS_PER_PAGE * pageIndex;
-
-		while (playerIndex < ((pageIndex + 1) * MAX_PLAYERS_PER_PAGE) &&
-			playerIndex < activePlayersCount)
-		{
-			User user = activeUsers.get(playerIndex);
-
-			panelBuilder.item(new PanelItemBuilder().
-				name(user.getName()).
-				icon(user.getName()).
-				clickHandler((panel, clicker, click, slot) -> {
-					this.player.closeInventory();
-
-					new BiomesPanel(
-						this.addon,
-						this.player,
-						user,
-						this.world,
-						"",
-						BiomesPanel.Mode.ADMIN);
-					return true;
-				}).build()
-			);
-
-			playerIndex++;
-		}
-
-		// Next next button
-		if (playerIndex < activeUsers.size())
-		{
-			final int nextPage = pageIndex + 1;
-
-			panelBuilder.item(MAX_PLAYERS_PER_PAGE + 8, new PanelItemBuilder().
-				name(this.player.getTranslation("biomes.gui.buttons.next")).
-				icon(new ItemStack(Material.SIGN)).
-				clickHandler((panel, clicker, click, slot) -> {
-					this.createPlayerListMenu(nextPage);
-					return true;
-				}).build());
-		}
-
-		// Previous button
-		if (playerIndex > MAX_PLAYERS_PER_PAGE)
-		{
-			final int previousPage = pageIndex - 1;
-
-			panelBuilder.item(MAX_PLAYERS_PER_PAGE, new PanelItemBuilder().
-				name(this.player.getTranslation("biomes.gui.buttons.previous")).
-				icon(new ItemStack(Material.SIGN)).
-				clickHandler((panel, clicker, click, slot) -> {
-					this.createPlayerListMenu(previousPage);
-					return true;
-				}).build());
-		}
 
 		panelBuilder.build();
 	}
